+++ src/constants/api-endpoints.js
@@ -1,13 +1,7 @@
-const topics = 'topics'
-const indexPage = 'index_page'
-const indexPageCategories = 'index_page_categories'
-const posts = 'posts'
-const users = 'users'
-
 export default {
-  topics,
-  indexPage,
-  indexPageCategories,
-  posts,
-  users,
+  topics: 'topics',
+  indexPage: 'index_page',
+  indexPageCategories: 'index_page_categories',
+  posts: 'posts',
+  users: 'users',
 }