+++ src/store/bind-actions-to-store.js
@@ -1,20 +1,5 @@
 import { bindActionCreators } from 'redux'
-import {
-  fetchIndexPageContent,
-  fetchCategoriesPostsOnIndexPage,
-} from '../actions/index-page'
-import {
-  fetchAFullPost,
-  fetchListedPosts,
-  fetchEditorPickedPosts,
-  fetchInfographicPostsOnIndexPage,
-  fetchPhotographyPostsOnIndexPage,
-} from '../actions/posts'
-import {
-  fetchAFullTopic,
-  fetchTopics,
-  fetchTopicsOnIndexPage,
-} from '../actions/topics'
+import actions from '../actions'
 
 /**
  * A store enhancer that will attach a property `actions` consisting of action creators bound to store.
@@ -27,21 +12,7 @@ import {
 export default function bindActionsToStore(prevStoreCreator) {
   return function(reducer, preloadedState) {
     const _store = prevStoreCreator(reducer, preloadedState)
-    _store.actions = bindActionCreators(
-      {
-        fetchIndexPageContent,
-        fetchCategoriesPostsOnIndexPage,
-        fetchAFullPost,
-        fetchListedPosts,
-        fetchEditorPickedPosts,
-        fetchInfographicPostsOnIndexPage,
-        fetchPhotographyPostsOnIndexPage,
-        fetchAFullTopic,
-        fetchTopics,
-        fetchTopicsOnIndexPage,
-      },
-      _store.dispatch
-    )
+    _store.actions = bindActionCreators(actions, _store.dispatch)
     return _store
   }
 }