+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### Unrelease
+## Unrelease
 
 ### 5.0.0
 
@@ -11,6 +11,10 @@
   - Remove `gulp` and use `makefile` instead
   - Update `.gitignore` and `.npmignore`
 - Add redux store
+- Move actions and reducers from `twreporter-react` to this package
+- Remove unused reducers and actions
+
+## Release
 
 ### 4.0.3